Changed singletons to be part of Amaster
Because singletons bad. tbh I don't understand the hate for singletons at all because used correctly it creates more flexible code due to access in global scope and only someone with the brain of a toddle couldnt understand such a simple concept. but whatever, muh patterns bad. broke colors though that is for tomorrow to fix